Fix `invalid function` errors when using native compilation
When native-compiled, instead of signature help for example, an error
is displayed in the echo area: `eldoc: Invalid function:
lsp-proxy--async-request`

`lsp-proxy--async-request` is a macro, and cannot be autoloaded by
declare-function. The simplest fix is just loading
`lsp-proxy-core`. At runtime it will be loaded anyway so there is no
benefit in not depending on it.

A similar issue is present for imenu and large file handling.